requirement.txt
fastapi
uvicorn
python-dotenv
langchain-groq
langchain-core
langchain-community
langchain-pinecone
langchain-huggingface
pinecone-client
sentence-transformers